server/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
SQLAlchemy==1.4.46
pydantic==1.10.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.4.0
stripe==4.34.0
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.5
typing-extensions==4.5.0